payment method done

+++ src/main/java/DTO/ItemDTO.java
@@ -3,7 +3,7 @@ package DTO;
 public class ItemDTO implements DTO {
     public final String name;
     public final double price;
-    public final int machineCode;
+    public final long machineCode;
 
     public ItemDTO(int machineCode, String name, double price) {
         this.name = name;
