+++ src/main/java/DTO/PaymentMethodDTO.java
@@ -2,9 +2,9 @@ package DTO;
 
 public class PaymentMethodDTO implements DTO {
     public final String name;
-    public final int machineCode;
+    public final long machineCode;
 
-    public PaymentMethodDTO(int machineCode, String name) {
+    public PaymentMethodDTO(long machineCode, String name) {
         this.name = name;
         this.machineCode = machineCode;
     }
