+++ src/main/java/DTO/TransactionDTO.java
@@ -3,7 +3,7 @@ package DTO;
 public class TransactionDTO implements DTO {
     public final int itemCode;
     public final int paymentMethod;
-    public final int machineCode;
+    public final long machineCode;
 
     public TransactionDTO(int machineCode, int itemCode, int paymentMethod) {
         this.machineCode = machineCode;
